calculatenow+result

+++ Home_GN/app/src/main/java/com/example/a59011178/home/cal/ResultActivity.java
@@ -29,7 +29,7 @@ public class ResultActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_result);
 
-        double result = getIntent().getExtras().getDouble("total");
+        float result = getIntent().getExtras().getFloat("total");
 
 
         TextView bahtresult = (TextView)findViewById(R.id.bahtresult);
